requirements.txt
Flask==0.10.1
flask-peewee==0.6.5
Flask-Security==1.6.9
Flask-OAuthlib==0.3.3
Flask-Script==0.6.2
PyMySQL==0.5
fitbit==0.0.2
gevent==0.13.8
bcrypt==1.0.2